// ui/src/a2a/schema.ts
export interface PartBase {
  /**
   * Extension metadata.
   */
  metadata?: {
    [key: string]: any;
  };
}

/**
 * File content with bytes.
 */
export interface FileWithBytes {
  /**
   * Optional name of the file.
   */
  name?: string;

  /**
   * Optional MIME type of the file content.
   */
  mimeType?: string;

  /**
   * File content encoded as a Base64 string.
   */
  bytes: string;
}

/**
 * File content with URI.
 */
export interface FileWithUri {
  /**
   * Optional name of the file.
   */
  name?: string;

  /**
   * Optional MIME type of the file content.
   */
  mimeType?: string;

  /**
   * URI pointing to the file content.
   */
  uri: string;
}

/**
 * Represents a text segment within parts.
 */
export interface TextPart extends PartBase {
  /**
   * Part type - text for TextParts.
   */
  kind: "text";

  /**
   * Text content.
   */
  text: string;
}

/**
 * Represents a File segment within parts.
 */
export interface FilePart extends PartBase {
  /**
   * Part type - file for FileParts.
   */
  kind: "file";

  /**
   * File content either as url or bytes.
   */
  file: FileWithBytes | FileWithUri;
}

/**
 * Represents a structured data segment within a message part.
 */
export interface DataPart extends PartBase {
  /**
   * Part type - data for DataParts.
   */
  kind: "data";

  /**
   * Structured data content.
   */
  data: {
    [key: string]: any;
  };
}

/**
 * Represents a single part of a multi-part message. Can be text, file, or data.
 */
export type Part = TextPart | FilePart | DataPart;

/**
 * Represents an artifact generated or used by a task, potentially composed of multiple parts.
 */
export interface Artifact {
  /**
   * Unique identifier for the artifact.
   */
  artifactId: string;

  /**
   * Optional name for the artifact.
   */
  name?: string;

  /**
   * Optional description for the artifact.
   */
  description?: string;

  /**
   * Artifact parts.
   */
  parts: Part[];

  /**
   * Extension metadata.
   */
  metadata?: {
    [key: string]: any;
  };

  /**
   * The URIs of extensions that are present or contributed to this Artifact.
   */
  extensions?: string[];
}